Rename command argument to `kubeClientQPS` and `kubeClientBurst`
Following the naming convention of Kubernetes Controller Manager
https://github.com/kubernetes/kubernetes/blob/3327bb367490d5ffa01eaa012253c478acbfdf52/staging/src/k8s.io/controller-manager/options/generic.go#L62-L63

+++ pkg/controllers/manager.go
@@ -28,10 +28,10 @@ func NewCacheClientBypassSecrets(cache cache.Cache, config *rest.Config, options
 	return cluster.DefaultNewClient(cache, config, options, append(uncachedObjects, &corev1.Secret{})...)
 }
 
-func GetConfigOrDieWithQPSAndBurst(qps int, burst int) *rest.Config {
+func GetConfigOrDieWithQPSAndBurst(qps float32, burst int) *rest.Config {
 	cfg := ctrl.GetConfigOrDie()
 	if qps > 0 {
-		cfg.QPS = float32(qps)
+		cfg.QPS = qps
 	}
 
 	if burst > 0 {
